+++ src/js/touch.js
@@ -27,7 +27,15 @@
 		data.axisY = data.axis === "y";
 
 		if (Formstone.support.pointer) {
-			touchAction(this, "none");
+			var action = "";
+			if (data.axisX) {
+				action += " pan-y";
+			}
+			if (data.axisY) {
+				action += " pan-x";
+			}
+
+			touchAction(this, action);
 
 			this.on(Events.pointerDown, data, onTouch);
 		} else {