+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.8.43
+
+* Fixed issue with IE touch-action in Touch
+
 ### 0.8.42
 
 * Fixed issue with destroying a Carousel